Prevent cold extrusions.
Why did I introduce a new status variable in the Extruder class when I simply
could have used the disable() method?

I need to disable extruder moves as soon as the first temperature reading is
done (if the temperature is too low). This means that the extruder is disabled
from the start until the min temperature is reached. As the enabled flag is
checked when GCODEs are received this would mean that all extruder moves that
fill up the queue at the beginning (while the system waits for the hotend to
reach temp.) would be dismissed - which would result in wrong operation when
nothing is wrong.

So I introduced a new status variable which then is used to prevent all
extruder moves upon GCODE execution, not validation (if the temperature is too
low).

Since wolfmanjm said that one could argue about if ON_HALT should be called
when the cold extrusion prevention kicks in I simply made it configurable.